refactor: :rainbow: 폴더 정리
