+++ iOS/IssueTracker/IssueTracker/02.LabelScene/View/LabelCell/LabelCellView.swift
@@ -54,13 +54,3 @@ extension LabelCellView: UICollectionViewRegisterable {
         return UINib(nibName: "LabelCellView", bundle: nil)
     }
 }
-
-extension UIColor {
-    var isDarkColor: Bool {
-        var r, g, b, a: CGFloat
-        (r, g, b, a) = (0, 0, 0, 0)
-        self.getRed(&r, green: &g, blue: &b, alpha: &a)
-        let lum = 0.2126 * r + 0.7152 * g + 0.0722 * b
-        return  lum < 0.50
-    }
-}
