+++ iOS/IssueTracker/IssueTracker/05.CustomUI/ImageLoadable.swift
