+++ iOS/IssueTracker/IssueTracker/05.CustomUI/LeftAlignedLabelCollection/LabelBadgeCellView.swift
